code/Retrieval/v1.0.0/requirements.txt
# Retrieval Pipeline API v1.0.0 Dependencies

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Utilities
python-multipart==0.0.20
